pyproject.toml
[tool.poetry]
name = "glassdoorcrawler"
version = "0.1.0"
description = "Crawler para coletar dados de vagas no Glassdoor"
authors = ["Lucas Romualdo Fernandes"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
beautifulsoup4 = "^4.13.0"
numpy = "^2.0.0"
openpyxl = "^3.1.0"
pandas = "^2.0.0"
progressbar2 = "^4.5.0"
requests = "^2.32.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
glassdoorcrawler = "glassdoorcrawler.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
